fix : project structure
